pyproject.toml
[project]
name = "ai-news-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "feedparser>=6.0.12",
    "groq>=1.2.0",
    "newsapi-python>=0.2.7",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "twilio>=9.10.5",
]
